org.eclipse.virgo.bundlor.ant/src/main/java/org/eclipse/virgo/bundlor/ant/ManifestTemplate.java
/*******************************************************************************
 * Copyright (c) 2008, 2010 VMware Inc.
 * All rights reserved. This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v1.0
 * which accompanies this distribution, and is available at
 * http://www.eclipse.org/legal/epl-v10.html
 *
 * Contributors:
 *   VMware Inc. - initial contribution
 *******************************************************************************/

package org.eclipse.virgo.bundlor.ant;

public final class ManifestTemplate {

    private volatile String template;

    public void addText(String template) {
        this.template = template;
    }

    String getTemplate() {
        return template;
    }
}
